requirements.txt
ago==0.0.95
async-generator==1.10
attrs==22.1.0
Automat==22.10.0
awscli==1.27.15
beautifulsoup4==4.9.3
botocore==1.29.15
bs4==0.0.1
cchardet==2.1.7
certifi==2022.9.14
cffi==1.15.1
chardet==4.0.0
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==38.0.3
cssselect==1.2.0
docutils==0.16
dotmap==1.3.30
elastic-transport==8.4.0
elasticsearch==8.5.2
exceptiongroup==1.0.4
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
googlesearch-python==1.1.0
h11==0.14.0
hjson==3.1.0
huggingface-hub==0.11.0
hurry.filesize==0.9
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jieba3k==0.35.1
Jinja2==3.0.1
jmespath==1.0.1
joblib==1.2.0
langdetect==1.0.9
lxml==4.9.1
MarkupSafe==2.0.1
news-please==1.5.22
newspaper3k==0.2.8
nltk==3.7
node==1.1
numpy==1.23.4
odict==1.9.0
outcome==1.2.0
packaging==21.3
parsel==1.7.0
Pillow==8.4.0
plac==1.3.5
plumber==1.7
Protego==0.2.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
PySocks==1.7.1
python-crfsuite==0.9.8
python-dateutil==2.8.2
PyYAML==5.4.1
queuelib==1.6.2
readability-lxml==0.8.1
regex==2022.10.31
requests==2.25.1
requests-file==1.5.1
rsa==4.7.2
s3transfer==0.6.0
scikit-learn==1.1.3
scipy==1.9.3
Scrapy==2.7.1
selenium==4.6.0
seqeval==1.2.2
service-identity==21.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tinysegmenter==0.3
tkdesigner==1.0.4
tldextract==3.4.0
tokenizers==0.13.2
torch==1.13.0
tqdm==4.64.1
transformers==4.24.0
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typing_extensions==4.4.0
underthesea==1.3.3
Unidecode==1.3.6
urllib3==1.26.14
w3lib==2.0.1
warcio==1.7.4
wsproto==1.2.0
zope.component==5.0.1
zope.deferredimport==4.4
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.4
zope.interface==5.5.2
zope.lifecycleevent==4.4
zope.proxy==4.6.1
